# BUG/PERF: DataFrame.isin lossy data conversion