--- a/pandas/tests/frame/methods/test_isin.py
+++ b/pandas/tests/frame/methods/test_isin.py
@@ -219,7 +219,7 @@ class TestDataFrameIsIn:
         tm.assert_frame_equal(result, expected)
 
     def test_isin_not_lossy(self):
-        # GH #####
+        # GH 53514
         val = 1666880195890293744
         df = DataFrame({"a": [val], "b": [1.0]})
         result = df.isin([val])