- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.1.0.rst` file if fixing a bug or adding a new feature.


edge case bug (lossy data conversion):
```
import pandas as pd

val = 1666880195890293744

df = pd.DataFrame({"a": [val], "b": [1.0]})

df.apply(lambda x: x.isin([val]))  # True, False
df.isin([val])                     # False, False 
```


perf (for EA dtypes):
```
import pandas as pd
import numpy as np

data = np.random.randint(0, 1000, (100_000, 2))
df = pd.DataFrame(data, dtype="int64[pyarrow]")

%timeit df.isin([1, 50])

# 25.6 ms ± 1.49 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  <- main
# 2.1 ms ± 265 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)   <- PR
```